Renamed files and moved `CodeGeneratorEvaluatorTool` to `tools/comment_resolver.py`
